pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Carlos Pardo",email = "carlos.pardo.b78@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.118.0,<0.119.0)",
    "sqlmodel (>=0.0.25,<0.0.26)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (==1.7.4)",
    "bcrypt (==4.0.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

